polybar: swap backlight and volume to fix spacing
This fixes spacing issues between the volume and battery modules which
both have colored backgrounds at times and not at other times. This
makes it impossible to keep the spacing between them consistent.

By inserting the backlight module between them, which never has
a colored background, it's possible to keep the spacing consistent.