dunst: overhaul styles to match rest of system
Remove gaps and replace them with a separator. This allows us to use
shadow on notifications as the notifications are presented in one visual
'window'.

Shadow allows us to remove the frame from notifications. But now we need
another way to indicate urgency. So give critical notifications a red
background and invert the text color.

This commit also removes the icon from critical notifications, and fixes
an issue where the font size would depend on whether the system was
started multi-head or not.