dunst: add colored highlights to notifications
This commit adds colored highlights to notifications, to draw attention
to them. It also changes the format to make them look a bit nicer.

This commit also removes the corner_radius property since we do not need
it as we are not using frames.

This commit also switches to softer foreground and background colors to
match the rest of the system.