nvim: combine file explorer keymaps
The <leader>E keymap is only used if <leader>e fails. This only happens
if we try to open the explorer in a buffer which does not represent
a file, for example the mini.starter buffer or nvim-dap buffers.

This commit updates <leader>e to check for failure. On failure, we open
an explorer in the current working directory, which is what <leader>E
did.

The <leader>E keymap is no longer needed, so we remove it.